requirements.txt
opencv-python
pillow
numpy
tensorboardx
argparse
tensorboard
scikit-image
scikit-learn
omegaconf
pytorch-lightning==1.2.10
torch==1.8.1
torchvision==0.8.2
torchvision
tqdm